requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==3.14.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
twilio==8.2.0
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.2
